Added toCNF method to convert formulas to CNF
The formula class now has a toCNF method which can convert any formula to conjunctive normal form. This will be useful for converting formulas to the Dimacs CNF format for use with Sat4j. However, the conversion has not been tested very well, so some bugs may not be found yet.